requirements.txt
aiogram>=3.4
aiohttp>=3.9
aiosqlite>=0.19
asyncpg>=0.29
python-dotenv>=1.0
